+++ src/app/(bottom-nav)/category/layout.tsx
@@ -1,7 +1,7 @@
-import QuickMenu from "@/components/layouts/categories/QuickMenu";
 import React from "react";
 import CategoryHeader from "@/components/headers/CategoryHeader";
 import BottomNavigationBar from "@/components/navbars/BottomNavigationBar";
+import QuickMenu from "@/components/layouts/category/QuickMenu";
 
 function layout({ children }: { children: React.ReactNode }) {
   return (
