+++ src/app/(bottom-nav)/my/layout.tsx
@@ -1,0 +1,5 @@
+function layout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>;
+}
+
+export default layout;
