+++ src/app/(bottom-nav)/my/page.tsx
@@ -1,0 +1,6 @@
+import { redirect } from "next/navigation";
+
+export default function page() {
+  redirect("/sign-in");
+  return <div>page</div>;
+}
