+++ src/app/(bottom-nav)/recent/layout.tsx
@@ -1,0 +1,12 @@
+import BottomNavigationBar from "@/components/navbars/BottomNavigationBar";
+
+function layout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      {children}
+      <BottomNavigationBar />
+    </>
+  );
+}
+
+export default layout;
