+++ src/app/category/[...path]/page.tsx
@@ -1,0 +1,3 @@
+export default function page() {
+  return <div>page</div>;
+}
